HydroSure_AI/requirements.txt
# --- Core Backend ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.8.2
pydantic-settings>=2.3.0
python-dotenv>=1.0.1

# --- LangGraph & AI ---
langgraph>=0.2.69
openai>=1.40.0

# --- Computer Vision & Image Processing ---
opencv-python>=4.10.0.84
scikit-image>=0.24.0
numpy>=1.26.4

# --- Database ---
pymongo>=4.7.2
